## Things I learned and Understand More - StringStream It is useful when you want to put any variables in a string without manually converting them. Also, StringStream is like an input that can convert a string to an int, double, etc.
- Getline Delimiter
I did not know that there was a delimiter feature in getline. Makes reading the text much easier.

- File Opening
By using the fstream libraries, I was able to open the CSV file with ease, and it is easy to store it in a string.

- Command-Line Parameters
The simplest way I can say this is that you can see the file where it is located, and also see what file it is using. That's why I was able to grab the CSV file from outside with ease.